+++ metta/rl/functions/advantage.py
@@ -1,0 +1,93 @@
+"""Advantage computation functions for Metta training."""
+
+from contextlib import nullcontext
+from typing import Union
+
+import einops
+import torch
+from pufferlib import _C  # noqa: F401 - Required for torch.ops.pufferlib
+from torch import Tensor
+
+from metta.rl import mps
+
+
+def compute_advantage(
+    values: Tensor,
+    rewards: Tensor,
+    dones: Tensor,
+    importance_sampling_ratio: Tensor,
+    advantages: Tensor,
+    gamma: float,
+    gae_lambda: float,
+    vtrace_rho_clip: float,
+    vtrace_c_clip: float,
+    device: Union[torch.device, str],
+) -> Tensor:
+    """CUDA kernel for puffer advantage with automatic CPU & MPS fallback.
+
+    This matches the trainer.py implementation exactly.
+    """
+    # Get correct device
+    device = torch.device(device) if isinstance(device, str) else device
+
+    # Move tensors to device and compute advantage
+    if str(device) == "mps":
+        return mps.advantage(
+            values, rewards, dones, importance_sampling_ratio, vtrace_rho_clip, vtrace_c_clip, gamma, gae_lambda, device
+        )
+
+    # CUDA implementation using custom kernel
+    tensors = [values, rewards, dones, importance_sampling_ratio, advantages]
+    tensors = [t.to(device) for t in tensors]
+    values, rewards, dones, importance_sampling_ratio, advantages = tensors
+
+    # Create context manager that only applies CUDA device context if needed
+    device_context = torch.cuda.device(device) if str(device).startswith("cuda") else nullcontext()
+    with device_context:
+        torch.ops.pufferlib.compute_puff_advantage(
+            values,
+            rewards,
+            dones,
+            importance_sampling_ratio,
+            advantages,
+            gamma,
+            gae_lambda,
+            vtrace_rho_clip,
+            vtrace_c_clip,
+        )
+
+    return advantages
+
+
+def normalize_advantage_distributed(adv: Tensor, norm_adv: bool = True) -> Tensor:
+    """Normalize advantages with distributed training support while preserving shape.
+
+    This matches the trainer.py implementation exactly.
+    """
+    if not norm_adv:
+        return adv
+
+    if torch.distributed.is_initialized():
+        # Compute local statistics
+        adv_flat = adv.view(-1)
+        local_sum = einops.rearrange(adv_flat.sum(), "-> 1")
+        local_sq_sum = einops.rearrange((adv_flat * adv_flat).sum(), "-> 1")
+        local_count = torch.tensor([adv_flat.numel()], dtype=adv.dtype, device=adv.device)
+
+        # Combine statistics for single all_reduce
+        stats = einops.rearrange([local_sum, local_sq_sum, local_count], "one float -> (float one)")
+        torch.distributed.all_reduce(stats, op=torch.distributed.ReduceOp.SUM)
+
+        # Extract global statistics
+        global_sum, global_sq_sum, global_count = stats[0], stats[1], stats[2]
+        global_mean = global_sum / global_count
+        global_var = (global_sq_sum / global_count) - (global_mean * global_mean)
+        global_std = torch.sqrt(global_var.clamp(min=1e-8))
+
+        # Normalize and reshape back
+        adv = (adv - global_mean) / (global_std + 1e-8)
+    else:
+        # Local normalization
+        adv = (adv - adv.mean()) / (adv.std() + 1e-8)
+
+    return adv
